+++ src/command_modules/azure-cli-vm/HISTORY.rst
@@ -5,6 +5,7 @@ Release History
 2.2.12
 ++++++
 * `vm/vmss create`: support cross tenant image referencing.
+* `vm diagnostics get-default-config --windows-os`: fix bug in windows os default config.
 * `vmss extension set`: expose `--provision-after-extensions`. It defines what extensions must be provisioned before the extension being set.
 
 2.2.11